+++ jumper/index.html
@@ -8,6 +8,8 @@
     <script src="components/sky-canvas.js"></script>
     <script src="https://unpkg.com/aframe-look-at-component@0.8.0/dist/aframe-look-at-component.min.js"></script>
     <link rel="stylesheet" href="style.css">
+
+    <script src="main.js" type="module"></script>
 </head>
 
 <body>
@@ -26,7 +28,7 @@
     </a-scene>
 
 
-    <script  src="./kaaro.js"></script>
+    <!-- <script  src="./kaaro.js"></script> -->
 </body>
 
 </html>
