+++ jumper/main.js
@@ -1,0 +1,4 @@
+import { listenForAllTheThingsTheUserSaysMostlyEntitiesAndShowThemOnTheViewer } from './kaaro.js';
+
+console.log("in main.js");
+listenForAllTheThingsTheUserSaysMostlyEntitiesAndShowThemOnTheViewer();
